+++ src/main/java/de/alpharogroup/file/modify/api/FileChangable.java
@@ -1,0 +1,7 @@
+package de.alpharogroup.file.modify.api;
+
+import java.util.function.BiFunction;
+
+public interface FileChangable extends BiFunction<Integer, String, String>
+{
+}