+++ src/main/java/de/alpharogroup/file/zip/ZipErrorCodes.java
@@ -2,5 +2,5 @@ package de.alpharogroup.file.zip;
 
 public enum ZipErrorCodes
 {
-	DIRECTORY_TO_ZIP_DOES_NOT_EXIST, ZIP_FILE_DOES_NOT_EXIST, IO_ERROR;
+	DIRECTORY_TO_ZIP_DOES_NOT_EXIST, IO_ERROR, ZIP_FILE_DOES_NOT_EXIST;
 }